ajout choix discution
